requirements.txt
grpcio
grpcio-tools
tinkoff-investments
pandas
matplotlib
ta